Update Bot.py
Отредактирована работа с командами

+++ Bot.py
@@ -133,33 +133,39 @@ write_data_to_file(ADMINLIST_FILE, "\n".join(admin_list))
 def handle_commands(message):
     # Проверка, что команда отправлена в личном сообщении и отправитель является администратором
     if str(message.from_user.id) in admin_ids:
-        # Создаем клавиатуру
-        keyboard = types.ReplyKeyboardMarkup(row_width=2)
-        button_texts = ["Добавить данные в BAN", "Добавить данные в WARNING", "Статистика", "Статус"]
-        for text in button_texts:
-            button = types.KeyboardButton(text=text)
-            keyboard.add(button)
-        # Отправляем сообщение с клавиатурой в личное сообщение администратору
-        try:
-            bot.send_message(message.from_user.id, "Выберите действие:", reply_markup=keyboard)
-        except ApiTelegramException as e:
-            if e.error_code == 403:
-                bot.send_message(message.chat.id,
-                                 "Для работы с ботом, пожалуйста, разрешите получение сообщений от ботов в настройках конфиденциальности Telegram.")
-            else:
-                bot.send_message(message.chat.id, f"Произошла ошибка: {e}")
-        # Удаление сообщения, если оно не отправлено в личном чате администратором
+        if str(message.text == "/start"):
+            # Создаем клавиатуру
+            keyboard = types.ReplyKeyboardMarkup(row_width=2)
+            button_texts = ["Добавить данные в BAN", "Добавить данные в WARNING", "Статистика", "Статус"]
+            for text in button_texts:
+                button = types.KeyboardButton(text=text)
+                keyboard.add(button)
+            # Отправляем сообщение с клавиатурой в личное сообщение администратору
+            try:
+                bot.send_message(message.from_user.id, "Выберите действие:", reply_markup=keyboard)
+            except ApiTelegramException as e:
+                if e.error_code == 403:
+                    bot.send_message(message.chat.id,
+                                     "Для работы с ботом, пожалуйста, разрешите получение сообщений от ботов в настройках конфиденциальности Telegram.")
+                else:
+                    bot.send_message(message.chat.id, f"Произошла ошибка: {e}")
+            # Удаление сообщения, если оно не отправлено в личном чате администратором
         bot.delete_message(message.chat.id, message.message_id)
-    else:
-        bot.delete_message(message.chat.id, message.message_id)
+
 
 # Обработчик для сообщений после выбора кнопки "Добавить данные в BAN"
 @bot.message_handler(func=lambda message: message.text == "Добавить данные в BAN")
 def add_to_ban_phrases(message):
-    # Запросить у пользователя текст для добавления в BAN
-    bot.send_message(message.chat.id, "Введите текст для добавления в BAN:")
-    # Добавить обработчик для следующего сообщения пользователя
-    bot.register_next_step_handler(message, process_ban_phrase)
+    if str(message.from_user.id) in admin_ids:
+        if str(message.text == "Добавить данные в BAN"):
+            try:
+                bot.send_message(message.from_user.id, "Введите текст для добавления в BAN:")
+                bot.delete_message(message.chat.id, message.message_id)
+                bot.register_next_step_handler(message, process_ban_phrase)
+            except Exception as e:
+                # Обработка ошибок
+                bot.reply_to(message, f"Произошла ошибка: {e}")
+                bot.delete_message(message.chat.id, message.message_id)
 
 # Функция для обработки текста, который нужно добавить в BAN
 def process_ban_phrase(message):
@@ -168,14 +174,22 @@ def process_ban_phrase(message):
     with open("banned_phrases.json", "a", encoding="utf-8") as file:
         file.write(new_phrase + "\n")
     # Отправить подтверждение администратору
-    bot.send_message(message.chat.id, f"Фраза '{new_phrase}' успешно добавлена в BAN.")
+    bot.send_message(message.from_user.id, f"Фраза '{new_phrase}' успешно добавлена в BAN.")
     banbanned_phrases = read_data_from_file(BANNED_PHRASES_FILE)
 
 # Обработчик для сообщений после выбора кнопки "Добавить данные в WARNING"
 @bot.message_handler(func=lambda message: message.text == "Добавить данные в WARNING")
 def add_to_warning_phrases(message):
-    bot.send_message(message.chat.id, "Введите текст для добавления в WARNING:")
-    bot.register_next_step_handler(message, process_warning_phrase)
+    if str(message.from_user.id) in admin_ids:
+        if str(message.text == "Добавить данные в WARNING"):
+            try:
+                bot.send_message(message.from_user.id, "Введите текст для добавления в WARNING:")
+                bot.delete_message(message.chat.id, message.message_id)
+                bot.register_next_step_handler(message, process_warning_phrase)
+            except Exception as e:
+                # Обработка ошибок
+                bot.reply_to(message, f"Произошла ошибка: {e}")
+                bot.delete_message(message.chat.id, message.message_id)
 
 # Функция для обработки текста, который нужно добавить в WARNING
 def process_warning_phrase(message):
@@ -189,9 +203,17 @@ def process_warning_phrase(message):
 # Обработчик для сообщений после выбора кнопки "Статистика"
 @bot.message_handler(func=lambda message: message.text == "Статистика")
 def handle_statistics(message):
-    bot.send_message(message.chat.id,
-                     "Введите интервал дат для вывода статистики в формате 'гггг-мм-дд гггг-мм-дд', например, '2024-02-01 2024-02-07':")
-    bot.register_next_step_handler(message, process_dates)
+    if str(message.from_user.id) in admin_ids:
+        if str(message.text == "Статистика"):
+            try:
+               bot.send_message(message.from_user.id,
+                             "Введите интервал дат для вывода статистики в формате 'гггг-мм-дд гггг-мм-дд', например, '2024-02-01 2024-02-07':")
+               bot.delete_message(message.chat.id, message.message_id)
+               bot.register_next_step_handler(message, process_dates)
+            except Exception as e:
+                # Обработка ошибок
+                bot.reply_to(message, f"Произошла ошибка: {e}")
+                bot.delete_message(message.chat.id, message.message_id)
 
 # Функция для обработки полученных дат
 def process_dates(message):
@@ -239,12 +261,17 @@ def count_events(file_path, file_bot_path, start_date, end_date):
 # Обработчик для сообщений после выбора кнопки "Статус"
 @bot.message_handler(func=lambda message: message.text == "Статус")
 def status_command(message):
-    try:
-        # Отправляем сообщение со статусом и временем запуска бота
-        bot.send_message(message.chat.id, f"Статус бота: онлайн\nВремя запуска бота: {bot_start_time}")
-    except Exception as e:
-        # Обработка ошибок
-        bot.reply_to(message, f"Произошла ошибка: {e}")
+    if str(message.from_user.id) in admin_ids:
+        if str(message.text == "Статус"):
+            try:
+                # Отправляем сообщение со статусом и временем запуска бота
+                bot.send_message(message.from_user.id, f"Статус бота: онлайн\nВремя запуска бота: {bot_start_time}")
+                bot.delete_message(message.chat.id, message.message_id)
+                bot.register_next_step_handler(message, process_dates)
+            except Exception as e:
+                # Обработка ошибок
+                bot.reply_to(message, f"Произошла ошибка: {e}")
+                bot.delete_message(message.chat.id, message.message_id)
 
 # Обработчик для всех сообщений
 @bot.message_handler(func=lambda message: True)
